API/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.2
blinker==1.4
cffi==1.15.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==4.1.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pkg_resources
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy
tomli==2.0.1
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
flask
flask_sqlalchemy
flask_login
flask_restful
flask_security
email_validator
flask_Security_too==4.1.4
flask_cors
redis
flask-caching
flask_caching
Flask-Caching
pandas
celery[redis]
celery
httplib2
jinja2
Flask-Mail
weasyprint
